Export to excel, Reading Overhaul
Overhauled how readings are stored.  Created two new objects that represent a reading from the gauge:
ReadingSet is a collection of multiple Readings.  When the program begins recording data from the gauge it will take the value and time from the gauge and store them as a Reading and add them to a ReadingSet.  When the gauge records the final reading after the delay it will finalize the ReadingSet and calculate some statistics like elapsed time and peak reading.  The reading sets are stored in a list in the main DPro app and that list is used to update the UI and also for exporting to excel.

Basic exporting is working.  The program will export to an excel sheet.  It will split the output into two sheets.  Sheet one will have the main statistics and final and peak values displayed.  The second sheet will have all the raw data from each ReadingSet (it just lists all of the stored Readings in each ReadingSet).  Exporting to excel is handled by Apache POI

Added a sound that plays when a reading is complete to alert the user that they can raise the gauge and start another reading